Refactor static methods around Paralyzer struct and expose Options
This change refactors existing Paralyzer functions into a single
`Paralyzer.DoContext` method which can take multiple options. The primary motivation
for this change is being able to use a combination of `ParalyzeWithLimit` and
`ParalyzeWithCancel` without passing a cancellation context into each
Paralyzable function. This new functionality is exposed as:

```go
NewParalyzer(WithConcurrencyLimit(32)).DoContext(ctx, funcs...)
```

The intention of the Options abstraction is so that new core functionality can
be exposed there and added to the `DoContext` method; however, this might not
be the correct interface for things like speculative execution.

Additionally, this PR makes the assumption that partial results and errors
will *not* be consumed if the Paralyzer panics and the caller recovers.